+++ docs/conf.py
@@ -42,7 +42,8 @@ extensions = [
     'rasabaster.card',
     'rasabaster.chatbubble',
     'rasabaster.runnable',
-    'rasabaster.copyable'
+    'rasabaster.copyable',
+    'rasabaster.apidoc',
 #    'numpydoc',
 ]
 
@@ -370,8 +371,11 @@ scv_greatest_tag = True
 
 
 def setup(sphinx):
+    sphinx.add_stylesheet('css/custom.css')
+
     try:
-        sys.path.insert(0, os.path.abspath('./util'))
+        utils_path = os.path.abspath(os.path.join(__file__, '..', 'util'))
+        sys.path.insert(0, utils_path)
         from StoryLexer import StoryLexer
         sphinx.add_lexer("story", StoryLexer())
     except ImportError: